+++ R/rtables_utils.R
@@ -353,28 +353,6 @@ ifneeded_split_col <- function(lyt, var, ...) {
   }
 }
 
-#' Create a Null Report
-#'
-#' @rdname report_null
-#' @aliases null_report
-#' @param tlg (`TableTree`) object.
-#' @param ... not used. Important to be used directly as post processing function.
-#'
-#' @export
-#'
-#' @return original `TableTree` or a null report if no observation are found in the table.
-report_null <- function(tlg, ...) {
-  assert_true(is.null(tlg) || rtables::is_rtable(tlg))
-
-  if (is.null(tlg) || nrow(tlg) == 0L) {
-    return(null_report)
-  }
-  if (count_children(tlg) == 0) {
-    return(null_report)
-  }
-  tlg
-}
-
 #' Count Children
 #'
 #' @keywords internal