+++ R/ttet01.R
@@ -153,7 +153,7 @@ ttet01_lyt <- function(arm_var,
       var_labels = timeunit,
       is_event = "IS_EVENT",
       control = control_survtp,
-      .labels = c("pt_at_risk" = render_safe("{Patient_label} remaining at risk")),
+      .labels = c("pt_at_risk" = render_safe("{Patients_label} remaining at risk")),
       ...
     )
 
@@ -177,10 +177,10 @@ ttet01_pre <- function(adam_db, dataset = "adtte",
       IS_NOT_EVENT = .data$CNSR == 1,
       EVNT1 = factor(
         case_when(
-          IS_EVENT == TRUE ~ render_safe("{Patient_label} with event (%)"),
-          IS_EVENT == FALSE ~ render_safe("{Patient_label} without event (%)")
+          IS_EVENT == TRUE ~ render_safe("{Patients_label} with event (%)"),
+          IS_EVENT == FALSE ~ render_safe("{Patients_label} without event (%)")
         ),
-        levels = render_safe(c("{Patient_label} with event (%)", "{Patient_label} without event (%)"))
+        levels = render_safe(c("{Patients_label} with event (%)", "{Patients_label} without event (%)"))
       ),
       EVNTDESC = factor(.data$EVNTDESC)
     )