+++ tests/testthat/test-dsl02.R
@@ -1,0 +1,16 @@
+# dsl02 ----
+
+test_that("dsl02 works with adsl dataset", {
+  expect_message(res <- run(dsl02, syn_data))
+  expect_snapshot(cat(export_as_txt(res[1:50, ], lpp = 100)))
+})
+
+# dsl02 functions ----
+
+test_that("dsl02 functions work as expected", {
+  expect_message(
+    res <- dsl02_pre(syn_data) %>%
+      dsl02_main()
+  )
+  expect_snapshot(cat(export_as_txt(res[1:50, ], lpp = 100)))
+})